+++ tests/RedisSortedSetsTest.php
@@ -22,114 +22,6 @@ class RedisSortedSetsTest extends TestCase
     }
 
 
-
-    /*
-     * ========================================================================
-     * zCard, zSize, zCount
-     * ========================================================================
-     */
-
-
-    /** @test */
-    public function redis_sorted_sets_zcount()
-    {
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $total = random_int(2, 10);
-        for ($i = 0; $i < $total; $i++) {
-            $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0 * ($i + 1), chr($i + 65)));
-        }
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(2, $this->redis->zCount($this->key, 1, 2));
-        $this->assertEquals($total, $this->redis->zCount($this->key, 1, $total));
-        $this->assertEquals(2, $this->redis->zCount($this->key, '-inf', 2));
-        $this->assertEquals($total, $this->redis->zCount($this->key, 1, '+inf'));
-        $this->assertEquals($total, $this->redis->zCount($this->key, '-inf', '+inf'));
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zsize_duplicate()
-    {
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        $this->assertEquals(0, $this->redis->zAdd($this->key, 2.0, 'A'));
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zSize($this->key));
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zsize_multiple()
-    {
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $total = random_int(1, 10);
-        for ($i = 0; $i < $total; $i++) {
-            $this->assertEquals(1, $this->redis->zAdd($this->key, 1.1 * $i, chr($i + 65)));
-        }
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals($total, $this->redis->zSize($this->key));
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zsize_single()
-    {
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zSize($this->key));
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zcard_duplicate()
-    {
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        $this->assertEquals(0, $this->redis->zAdd($this->key, 2.0, 'A'));
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zCard($this->key));
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zcard_multiple()
-    {
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $total = random_int(1, 10);
-        for ($i = 0; $i < $total; $i++) {
-            $this->assertEquals(1, $this->redis->zAdd($this->key, 1.1 * $i, chr($i + 65)));
-        }
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals($total, $this->redis->zCard($this->key));
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zcard_single()
-    {
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zCard($this->key));
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-    }
-
-
     /*
      * ========================================================================
      * zAdd
